Delete sgl-kernel AOT `bmm_fp8`, use `flashinfer.bmm_fp8` (#31202)
